pyproject.toml
[tool.poetry]
name = "document-pipeline"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
weaviate-client = "^4.14.1"
typer = "^0.15.2"
pydantic-settings = "^2.9.1"
tenacity = "^9.1.2"
yt-dlp = "^2025.3.31"
langchain = ">=0.3,<0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"